requirements.txt
pandas==1.2.4
python-binance==1.0.12
requests==2.25.1
sqlalchemy==1.4.17
psycopg2==2.8.6